refactor: Profile페이지 Profile폴더로 묶음
