+++ breaking-front/src/pages/Profile/Profile/Profile.js
@@ -2,18 +2,18 @@ import Line from 'components/Line/Line';
 import ProfileImage from 'components/ProfileImage/ProfileImage';
 import { ProfileSkeleton } from 'components/Skeleton/Skeleton';
 import Tabs from 'components/Tabs/Tabs';
-import useProfile from 'pages/Profile/hooks/queries/useProfile';
-import useProfileBookmarkedPost from 'pages/Profile/hooks/queries/useProfileBookmarkedPost';
-import useProfileBoughtPost from 'pages/Profile/hooks/queries/useProfileBoughtPost';
-import useProfileWrittenPost from 'pages/Profile/hooks/queries/useProfileWrittenPost';
+import useProfile from 'pages/Profile/Profile/hooks/queries/useProfile';
+import useProfileBookmarkedPost from 'pages/Profile/Profile/hooks/queries/useProfileBookmarkedPost';
+import useProfileBoughtPost from 'pages/Profile/Profile/hooks/queries/useProfileBoughtPost';
+import useProfileWrittenPost from 'pages/Profile/Profile/hooks/queries/useProfileWrittenPost';
 import useCheckMyPage from 'hooks/useCheckMyPage';
-import * as Style from 'pages/Profile/Profile.styles';
+import * as Style from 'pages/Profile/Profile/Profile.styles';
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import ImageUrlConverter from 'utils/ImageUrlConverter';
 import FollowButton from 'components/FollowButton/FollowButton';
-import ProfileFollowModal from 'pages/Profile/components/ProfileFollowModal/ProfileFollowModal';
-import ProfileTabPanel from 'pages/Profile/components/ProfileTabPanel/ProfileTabPanel';
+import ProfileFollowModal from 'pages/Profile/Profile/components/ProfileFollowModal/ProfileFollowModal';
+import ProfileTabPanel from 'pages/Profile/Profile/components/ProfileTabPanel/ProfileTabPanel';
 import numberFormatter from 'utils/numberFormatter';
 import useFollow from 'hooks/mutations/useFollow';
 import useUnFollow from 'hooks/mutations/useUnFollow';
