+++ breaking-front/src/pages/Profile/Profile/hooks/queries/useProfileBoughtPost.js
